requirements.txt
flask~=1.1.2
line-bot-sdk
configparser~=5.0.0
django
gunicorn
django-heroku
requests
werkzeug>=0.15.4

argparse~=1.4.0
pytz~=2020.1
moviepy>=1.0.3
git+https://github.com/pytube/pytube

typing-extensions
flask_sqlalchemy~=2.5.1
flask_migrate~=3.1.0
sqlalchemy>=1.4.0
psycopg2-binary~=2.9.1
schedule~=1.1.0